# Switch sequence representation to u64