--- a/src/cluster.rs
+++ b/src/cluster.rs
@@ -41,12 +41,10 @@ pub fn cluster(
         //let query_vec = seq.iter().map(|c| encode_single(*c)).collect::<Vec<_>>();
         let query_vec =
             SeqEncodingLength::from_bytes(record_unwrapped.id(), &record_unwrapped.seq());
-        // Skip if sequence has already been seen
-        // TODO: Only hash once here
-        if seen_sequences.contains(&query_vec.encoding.0) {
+
+        // Skip if sequence has already been seen (in which case insert returns false)
+        if !seen_sequences.insert(query_vec.encoding.0.clone()) {
             continue;
-        } else {
-            seen_sequences.insert(query_vec.encoding.0.clone());
         }
 
         // Get distances
@@ -71,7 +69,7 @@ pub fn cluster(
         } else {
             // If distance >= max_divergence then add to new centroid
             assigned_centroid = centroids.windows.len();
-            centroids.push_encoding(query_vec.clone());
+            centroids.push_encoding(query_vec);
             distances.push(0); // Adding another entry so that distances.len() == centroids.windows.len()
         }
         debug!("Assigned centroid: {}", assigned_centroid);