… tests/data/random_3_2.fna.smafadb
… tests/data/random_3_2_one_repeated.fna.smafadb