This PR is a rollup of several changes - you can review the individual commits (which are each fully functional and passes tests) if you want.

The major changes are
* Sequences are now encoded in u64, (as originally discussed in #46) . This implies the database format has changed and is also now more compact.
* Smafa now handles lowercase DNA sequences as well
* Slightly improved error message when a sequence cannot be encoded as DNA
* Smafa now verifies that sequences have the same length

Here is a benchmark of aligning 50,000 random DNA sequences against 1000 reference sequences:

Before
```
Benchmark 1: ./target/release/smafa query -q ~/Downloads/smafa/query.fna -d ~/Downloads/smafa/subject.fna.3cfb > /dev/null 2> /dev/null
  Time (mean ± σ):      6.944 s ±  0.069 s    [User: 6.884 s, System: 0.022 s]
  Range (min … max):    6.833 s …  7.010 s    5 runs
```

After:
```
Benchmark 1: ./target/release/smafa query -q ~/Downloads/smafa/query.fna -d ~/Downloads/smafa/subject.fna.0c41 > /dev/null 2> /dev/null
  Time (mean ± σ):     272.4 ms ±   1.3 ms    [User: 256.3 ms, System: 14.3 ms]
  Range (min … max):   271.1 ms … 274.5 ms    5 runs
```

Closes #46