@wwood 2024-03-01
Hey, this looks awesome. I didn't realise that each bool was actually taking up 8 bits.

Right now you are only using 5 bits of each 8 though, right? I wonder if it could be pushed further, maybe using a bit field?

I'll do a proper review when I'm back in the office, but looks good so far.
Thanks again.