@wwood 2024-03-01
Hmm, I think what you are saying about the masking only applies if we are using `==` between the bases, correct? 

If we have a proper `xor` between an array of bits (as opposed to bools) then we shouldn't need to mask, we can just use `distance = sum(xor) / 2`. This was the original intention, but I had naively assumed that an array of bools would be implemented as an array of bits, not an array of bytes.

A way forward is to use this crate's method, which is apparently not in the official documentation, but supposedly actually works.
https://github.com/contain-rs/bit-vec/blob/master/src/lib.rs#L782

WDYT? There's other useful things in this PR of course, which are probably fine to merge, but I'd like your opinion first.