@jakobnissen 2024-03-04
That sounds like a good idea. I'm a little busy the next days, but at the end of the week I can look at it again.
We could pack 12 nucleotides in the 12*5=60 lower bits of each `u64`, then store the sequences as `Vec<u64>`.
This would allow us to use `u64::count_ones`, which is more efficient than doing that on smaller integers like u8.
We could of course pack the nucleotides densely, using all bits of the `u64`, but this is annoying to do and only gets a small density increase.

Some preliminary microbenchmarking in Julia says that is about 2x faster than what the code does now, so it might well be worth it!