@wwood 2024-03-06
This seems pretty awesome, and so wil be a significant speedup for smafa and also SingleM. Nice. I would imagine that the fact they are random won't make any difference, but we'll see. I think aligning 1000 queries against 50k references would be more realistic, but no need to do further bunchmarking here.

I think everything seems good now, but it would probably be ideal if the behaviour was a bit nicer when the DB is in the old format. I'm happy to implement that but snowed under for a few days.

Is there some standard way of dealing with changing DB formats that you know of? Right now I was just thinking of attempting to deserialise with the newest format, and if that fails trying the old format. If the old one works then just error out with a helpful message - I don't think we need to maintain backwards compatibility.

Thanks, ben