@jakobnissen 2024-03-06
> Is there some standard way of dealing with changing DB formats that you know of?
The way I normally do it is to open the file, read the first few bytes to determine the database version and then use a match statement to pick the correct parser.