@wwood 2024-03-10
Thanks, done. I realised also that the version number wasn't updated in this PR. That's now done so the current DB version is now 2.